Modified UNVMe driver to use UIO device and no-IOMMU VFIO
This is the main transition away from an IOMMU-based driver to something that
can work with a statically-mapped buffer defined with a UIO device, alongside
VFIO running in no-IOMMU mode for access to PCIe.

This change allows the driver to run on embedded devices without an IOMMU, but
with the flexibility to allocate a chunk of memory to UNVMe through the device
tree, such as the Xilinx Zynq UltraScale+ MPSoC.

The implementation is currently quite restricted, requiring a UIO driver to be
loaded at /dev/uio0, with its first memory map a region of DMA-able memory of
size 1GiB and address 0x40000000.